requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colored==2.2.3
colorlog==6.7.0
distlib==0.3.7
filelock==3.12.2
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
multidict==6.0.4
outcome==1.3.0.post0
pipenv==2023.7.23
platformdirs==3.10.0
PySocks==1.7.1
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.0.4
virtualenv==20.24.2
virtualenv-clone==0.5.7
wsproto==1.2.0
yarl==1.9.2
